Add Github Action, adapt to linux
This adds the Github Action `acceptance` workflow for running acceptance
tests. Also, some of the initialization was adapted to run properly on
other linux environments.

One of the most noticeable changes is that `TESTER_ADDRESS` is now set
when a new aceptadora is created.

We've also adapted config matchers to be more generic and compsable,
which is a breaking change but we didn't release any version yet.