v0.2.0 ImagePuller as aceptadora.New() param
Instead of instantiating ImagePuller for each aceptadora, we now can
reuse a global one (for the test suite). This will avoid pulling same
image for each test where a new aceptadora is created every time.

We need this as docker repository is now imposing rate limit and
introducing latency on docker pulls, and we're seeing times of up to
2.5s to just check that we have the latest image.

This is a breaking change.